tests moved to tests folder; resource folder created
